Fix printf format for st_size

+++ xml.c
@@ -63,7 +63,8 @@ int config_lookup_host(struct openconnect_info *vpninfo, const char *host)
 
 	xmlfile = malloc(st.st_size);
 	if (!xmlfile) {
-		fprintf(stderr, "Could not allocate %zd bytes for XML config file\n", st.st_size);
+		fprintf(stderr, "Could not allocate %lu bytes for XML config file\n",
+			(unsigned long)st.st_size);
 		close(fd);
 		return -1;
 	}
